refactor: convert module to typescript
